Move Rate us core dialog to main package and get rid of common module.
