+++ app/src/main/java/wordtextcounter/details/main/util/RateUsHelper.kt
@@ -1,0 +1,51 @@
+package wordtextcounter.details.main.util
+
+import android.app.Activity
+import android.app.Dialog
+import android.content.Intent
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
+import android.net.Uri
+import android.view.LayoutInflater
+import android.view.Window
+import android.widget.Button
+import android.widget.ImageView
+import android.widget.Toast
+import com.example.rateus.R
+import com.example.rateus.RateusCore.onCancelSelected
+import com.example.rateus.RateusCore.onYesSelected
+import com.example.rateus.RateusCore.shouldShowRateUsDialog
+
+object RateUsHelper {
+  fun showRateUsDialog(activity: Activity) {
+    if (shouldShowRateUsDialog(activity)) {
+      val cView = LayoutInflater.from(activity)
+          .inflate(R.layout.rate_us_dialog, null)
+
+
+      val dialog = Dialog(activity)
+      dialog.requestWindowFeature(Window.FEATURE_NO_TITLE)
+      dialog.setContentView(cView)
+      dialog.window.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
+      val btClose = dialog.findViewById<ImageView>(R.id.ivCross)
+      val btYes = dialog.findViewById<Button>(R.id.btnSave)
+      btClose.setOnClickListener {
+        onCancelSelected(activity)
+        dialog.dismiss()
+      }
+
+      btYes.setOnClickListener {
+        onYesSelected(activity)
+        try {
+          val viewIntent = Intent(Intent.ACTION_VIEW,
+              Uri.parse("https://play.google.com/store/apps/details?id=wordtextcounter.details.main"))
+          activity.startActivity(viewIntent)
+        } catch (e: Exception) {
+          Toast.makeText(activity, "Oops! We were not able to reach play store. Sorry for that!", Toast.LENGTH_LONG).show()
+        }
+      }
+
+      dialog.show()
+    }
+  }
+}
