+++ app/src/main/res/drawable/ic_clear_black_24dp.xml
