+++ app/src/main/res/drawable/ic_rate_star.xml
