+++ app/src/main/res/layout/rate_us_dialog.xml
