+++ app/src/main/res/values-ar/strings.xml
