+++ app/src/main/res/values-de/strings.xml
