+++ app/src/main/res/values-gd/strings.xml
