+++ app/src/main/res/values-gu/strings.xml
