+++ app/src/main/res/values-hu/strings.xml
