+++ app/src/main/res/values-kn/strings.xml
