+++ app/src/main/res/values-ko/strings.xml
