+++ app/src/main/res/values-la/strings.xml
