+++ app/src/main/res/values-lv/strings.xml
