+++ app/src/main/res/values-ru/strings.xml
