+++ app/src/main/res/values-tr/strings.xml
