+++ app/src/main/res/values/palette.xml
