+++ app/src/main/res/values/strings.xml
@@ -43,6 +43,10 @@
   <string name="no_saved_notes">¯\\_(ツ)_/¯ \n\nNothing here!</string>
 
 
+  <!-- Rate us strings -->
+  <string name="rate_us_text">If you liked our app, would you please rate us on play store?</string>
+  <string name="rate_us_yes">Sure! Let\'s do it</string>
+
   <!-- Analytics consent strings -->
 
   <string name="analytics_consent" translatable="false">We wish to monitor behaviour of the app to understand how to make it better. is that okay?</string>
