+++ rateus/build.gradle
@@ -32,7 +32,6 @@ dependencies {
     api 'androidx.core:core-ktx:0.3'
     api "com.intuit.sdp:sdp-android:1.0.5"
     api "com.intuit.ssp:ssp-android:1.0.5"
-    implementation project(':common')
 }
 repositories {
     mavenCentral()
