--- rateus/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<resources>
-  <string name="rate_us_text">If you liked our app, would you please rate us on play store?</string>
-  <string name="rate_us_yes">Sure! Let\'s do it</string>
-</resources>
